Move styles to a stylesheet and other changes
