+++ examples/parlai_chat_task_demo/webapp/src/components/ConnectionIndicator.jsx
@@ -1,3 +1,11 @@
+/*
+ * Copyright (c) 2017-present, Facebook, Inc.
+ * All rights reserved.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
 import React from "react";
 import { CONNECTION_STATUS } from "mephisto-task";
