+++ examples/parlai_chat_task_demo/webapp/src/components/TextResponse.jsx
@@ -1,0 +1,118 @@
+/*
+ * Copyright (c) 2017-present, Facebook, Inc.
+ * All rights reserved.
+ * This source code is licensed under the BSD-style license found in the
+ * LICENSE file in the root directory of this source tree. An additional grant
+ * of patent rights can be found in the PATENTS file in the same directory.
+ */
+
+import React from "react";
+import { FormControl, Button } from "react-bootstrap";
+import $ from "jquery";
+
+class TextResponse extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = { textval: "", sending: false };
+  }
+
+  componentDidUpdate(prevProps, prevState, snapshot) {
+    // Only change in the active status of this component should cause a
+    // focus event. Not having this would make the focus occur on every
+    // state update (including things like volume changes)
+    if (this.props.active && !prevProps.active) {
+      $("input#id_text_input").focus();
+    }
+    // this.props.onInputResize();
+  }
+
+  tryMessageSend() {
+    if (this.state.textval !== "" && this.props.active && !this.state.sending) {
+      this.setState({ sending: true });
+      this.props
+        .onMessageSend({ text: this.state.textval, task_data: {} })
+        .then(() => this.setState({ textval: "", sending: false }));
+    }
+  }
+
+  handleKeyPress(e) {
+    if (e.key === "Enter") {
+      this.tryMessageSend();
+      e.stopPropagation();
+      e.nativeEvent.stopImmediatePropagation();
+    }
+  }
+
+  render() {
+    // TODO maybe move to CSS?
+    let pane_style = {
+      paddingLeft: "25px",
+      paddingTop: "20px",
+      paddingBottom: "20px",
+      paddingRight: "25px",
+      float: "left",
+      width: "100%",
+    };
+    let input_style = {
+      height: "50px",
+      width: "100%",
+      display: "block",
+      float: "left",
+    };
+    let submit_style = {
+      width: "100px",
+      height: "100%",
+      fontSize: "16px",
+      float: "left",
+      marginLeft: "10px",
+      padding: "0px",
+    };
+
+    let text_input = (
+      <FormControl
+        type="text"
+        id="id_text_input"
+        style={{
+          width: "80%",
+          height: "100%",
+          float: "left",
+          fontSize: "16px",
+        }}
+        value={this.state.textval}
+        placeholder="Please enter here..."
+        onKeyPress={(e) => this.handleKeyPress(e)}
+        onChange={(e) => this.setState({ textval: e.target.value })}
+        disabled={!this.props.active || this.state.sending}
+      />
+    );
+
+    let submit_button = (
+      <Button
+        className="btn btn-primary"
+        style={submit_style}
+        id="id_send_msg_button"
+        disabled={
+          this.state.textval === "" || !this.props.active || this.state.sending
+        }
+        onClick={() => this.tryMessageSend()}
+      >
+        Send
+      </Button>
+    );
+
+    return (
+      <div
+        id="response-type-text-input"
+        className="response-type-module"
+        style={pane_style}
+      >
+        <div style={input_style}>
+          {text_input}
+          {submit_button}
+        </div>
+      </div>
+    );
+  }
+}
+
+export default TextResponse;
