+++ examples/parlai_chat_task_demo/webapp/src/css/style.css
@@ -12,8 +12,38 @@
   display: flex;
 }
 
-#right-top-pane {
+.right-top-pane {
   flex: 1;
+  width: 100%;
+  position: relative;
+}
+
+.right-pane {
+  min-height: 100%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+}
+
+.message-pane-segment {
+  width: 100%;
+  padding-top: 60px;
+  padding-left: 20px;
+  padding-right: 20px;
+  padding-bottom: 20px;
+  overflow-y: scroll;
+}
+
+.preview-screen,
+.left-pane {
+  background-color: #dff0d8;
+  padding: 30px;
+  overflow: auto;
+}
+
+.right-bottom-pane {
+  width: 100%;
+  background-color: #eee;
 }
 
 .connection-indicator {
@@ -26,4 +56,23 @@
 .hourglass-container {
   margin-right: 5px;
   display: inline;
+}
+
+.chat-status-bar {
+  position: absolute;
+  background-color: #eeeeee;
+  border-color: #e7e7e7;
+  height: 46px;
+  top: 0;
+  border-width: 0 0 1px;
+  border-radius: 0;
+  right: 0;
+  left: 0;
+  z-index: 1030;
+  padding: 5px;
+}
+
+.chat-message {
+  float: left;
+  display: table;
 }
